+++ core/scripts/generate-contentful-schema.ts
@@ -4,7 +4,235 @@ import { readFileSync, writeFileSync } from 'fs';
 import { resolve } from 'path';
 import { z } from 'zod';
 
-import { generateContentfulSchemas } from '../contentful/schema-generator';
+// ================= Contentful Schema Types & Generators =================
+
+// Types for Contentful content type definitions
+interface ContentfulField {
+  id: string;
+  name: string;
+  type: string;
+  localized: boolean;
+  required: boolean;
+  validations: any[];
+  items?: {
+    type: string;
+    validations: any[];
+    linkType?: string;
+  };
+  linkType?: string;
+}
+
+interface ContentfulContentType {
+  sys: {
+    id: string;
+  };
+  name: string;
+  fields: ContentfulField[];
+}
+
+interface ContentfulSchema {
+  contentTypes: ContentfulContentType[];
+}
+
+// Helper functions for unpublished reference filtering
+function filterUnpublished<T extends { sys?: { publishedVersion?: number | null } }>(
+  arr: T[],
+): T[] {
+  return arr.filter((item) => item?.sys?.publishedVersion);
+}
+function singleReferenceOrNull<T extends { sys?: { publishedVersion?: number | null } }>(
+  val: T | null | undefined,
+): T | null {
+  return val?.sys?.publishedVersion ? val : null;
+}
+
+// Helper to create base field schema based on Contentful field type
+function createBaseFieldSchema(field: Partial<ContentfulField>): z.ZodTypeAny {
+  switch (field.type) {
+    case 'Symbol':
+    case 'Text':
+      return z.string();
+    case 'Integer':
+      return z.number().int();
+    case 'Number':
+      return z.number();
+    case 'Boolean':
+      return z.boolean();
+    case 'Date':
+      return z.string().datetime();
+    case 'Object':
+      return z.record(z.unknown());
+    case 'RichText': {
+      const simplifiedRichTextNode: z.ZodType<any> = z.lazy(() =>
+        z.object({
+          nodeType: z.string(),
+          data: z.record(z.unknown()),
+          content: z.array(z.any()).optional(),
+          marks: z.array(z.any()).optional(),
+          value: z.string().optional(),
+        }),
+      );
+      const simplifiedRichTextSchema = z.object({
+        nodeType: z.literal(BLOCKS.DOCUMENT),
+        data: z.record(z.unknown()),
+        content: z.array(simplifiedRichTextNode),
+      });
+      return simplifiedRichTextSchema;
+    }
+    case 'Array':
+      if (field.items) {
+        const itemSchema =
+          field.items.type === 'Link'
+            ? createLinkSchema(field.items.linkType)
+            : createBaseFieldSchema({ type: field.items.type });
+        // If this is an array of references, add the transform
+        if (field.items.type === 'Link') {
+          return z.array(itemSchema).transform(filterUnpublished);
+        }
+        return z.array(itemSchema);
+      }
+      return z.array(z.unknown());
+    case 'Link': {
+      // For single references, return null if unpublished
+      const linkSchema = createLinkSchema(field.linkType);
+      return linkSchema.transform(singleReferenceOrNull);
+    }
+    default:
+      return z.unknown();
+  }
+}
+
+// Helper to create link schema
+function createLinkSchema(linkType?: string): z.ZodTypeAny {
+  const detailedSysSchema = z.object({
+    space: z.object({
+      sys: z.object({
+        type: z.literal('Link'),
+        linkType: z.literal('Space'),
+        id: z.string(),
+      }),
+    }),
+    id: z.string(),
+    type: z.union([z.literal('Entry'), z.literal('Asset')]),
+    createdAt: z.string().datetime(),
+    updatedAt: z.string().datetime(),
+    environment: z.object({
+      sys: z.object({
+        id: z.string(),
+        type: z.literal('Link'),
+        linkType: z.literal('Environment'),
+      }),
+    }),
+    publishedVersion: z.number().optional(),
+    revision: z.number(),
+    locale: z.string().optional(),
+    contentType: z
+      .object({
+        sys: z.object({
+          type: z.literal('Link'),
+          linkType: z.literal('ContentType'),
+          id: z.string(),
+        }),
+      })
+      .optional(),
+  });
+  const metadataSchema = z.object({
+    tags: z.array(z.unknown()),
+    concepts: z.array(z.unknown()),
+  });
+  const assetFieldsSchema = z.object({
+    title: z.string().optional(),
+    description: z.string().optional(),
+    file: z.object({
+      url: z.string(),
+      details: z.object({
+        size: z.number(),
+        image: z
+          .object({
+            width: z.number(),
+            height: z.number(),
+          })
+          .optional(),
+      }),
+      fileName: z.string(),
+      contentType: z.string(),
+    }),
+  });
+  const linkedItemBaseSchema = z.object({
+    metadata: metadataSchema,
+    sys: detailedSysSchema,
+  });
+  if (linkType === 'Asset') {
+    return linkedItemBaseSchema.extend({
+      sys: detailedSysSchema.extend({
+        type: z.literal('Asset'),
+        contentType: z.undefined().optional(),
+      }),
+      fields: assetFieldsSchema,
+    });
+  } else if (linkType === 'Entry') {
+    return linkedItemBaseSchema.extend({
+      sys: detailedSysSchema.extend({
+        type: z.literal('Entry'),
+        contentType: z.object({
+          sys: z.object({
+            type: z.literal('Link'),
+            linkType: z.literal('ContentType'),
+            id: z.string(),
+          }),
+        }),
+      }),
+      fields: z.record(z.unknown()),
+    });
+  } else {
+    return z.union([
+      linkedItemBaseSchema.extend({
+        sys: detailedSysSchema.extend({
+          type: z.literal('Asset'),
+          contentType: z.undefined().optional(),
+        }),
+        fields: assetFieldsSchema,
+      }),
+      linkedItemBaseSchema.extend({
+        sys: detailedSysSchema.extend({
+          type: z.literal('Entry'),
+          contentType: z
+            .object({
+              sys: z.object({
+                type: z.literal('Link'),
+                linkType: z.literal('ContentType'),
+                id: z.string(),
+              }),
+            })
+            .optional(),
+        }),
+        fields: z.record(z.unknown()),
+      }),
+    ]);
+  }
+}
+
+function createContentTypeSchema(contentType: ContentfulContentType): Record<string, z.ZodTypeAny> {
+  const fieldSchemas: Record<string, z.ZodTypeAny> = {};
+  for (const field of contentType.fields) {
+    let fieldSchema = createBaseFieldSchema(field);
+    if (!field.required) {
+      fieldSchema = fieldSchema.optional();
+    }
+    fieldSchemas[field.id] = fieldSchema;
+  }
+  return fieldSchemas;
+}
+
+function generateContentfulSchemas(
+  contentfulSchema: ContentfulSchema,
+): Record<string, Record<string, z.ZodTypeAny>> {
+  const schemas: Record<string, Record<string, z.ZodTypeAny>> = {};
+  for (const contentType of contentfulSchema.contentTypes) {
+    schemas[contentType.sys.id] = createContentTypeSchema(contentType);
+  }
+  return schemas;
+}
 
 // Read the schema.json file
 const contentfulJsonPath = resolve(__dirname, '../contentful/schema.json');
@@ -16,6 +244,16 @@ const fieldSchemasMap = generateContentfulSchemas(contentfulSchema);
 // Name for the known recursive Rich Text schema
 const RICH_TEXT_SCHEMA_NAME = 'RichTextNodeSchema';
 
+// ========================================
+// Helper functions for transforms
+// ========================================
+const filterUnpublishedHelper = `// Utility to filter unpublished references in arrays\nfunction filterUnpublished<T extends { sys?: { publishedVersion?: number | null } }>(arr: T[]): T[] {\n  return arr.filter((item) => item?.sys?.publishedVersion);\n}`;
+const singleReferenceTransformHelper = `// Utility to return null for unpublished single references\nfunction singleReferenceOrNull<T extends { sys?: { publishedVersion?: number | null } }>(val: T | null | undefined): T | null {\n  return val?.sys?.publishedVersion ? val : null;\n}`;
+
+// --- Trackers for helper emission ---
+let usesFilterUnpublished = false;
+let usesSingleReferenceTransform = false;
+
 // Helper to convert a field schema to a string representation (Simplified for recursion)
 function zodToString(schema: any, indentationLevel = 0): string {
   if (!schema?._def) return 'z.unknown()';
@@ -23,14 +261,27 @@ function zodToString(schema: any, indentationLevel = 0): string {
   const indent = '  '.repeat(indentationLevel);
   const nextIndent = '  '.repeat(indentationLevel + 1);
 
-  // --- Type Handling (Simplified Lazy Handling) ---
   switch (schema._def.typeName) {
     case 'ZodLazy':
-      // Directly return the predefined name. Assumes this is the Rich Text schema.
-      // This prevents recursion during string generation.
       return RICH_TEXT_SCHEMA_NAME;
-    case 'ZodEffects':
+    case 'ZodEffects': {
+      const effect = schema._def.effect;
+      if (effect && effect.type === 'transform') {
+        const fnStr = effect.transform.toString();
+        if (
+          fnStr.includes('filterUnpublished') ||
+          (fnStr.includes('publishedVersion') && fnStr.includes('filter'))
+        ) {
+          usesFilterUnpublished = true;
+          return `${zodToString(schema._def.schema, indentationLevel)}.transform(filterUnpublished)`;
+        }
+        if (fnStr.includes('publishedVersion') && fnStr.includes('val ?')) {
+          usesSingleReferenceTransform = true;
+          return `${zodToString(schema._def.schema, indentationLevel)}.transform(singleReferenceOrNull)`;
+        }
+      }
       return zodToString(schema._def.schema, indentationLevel);
+    }
     case 'ZodString':
       return schema._def.checks?.some((c: any) => c.kind === 'datetime')
         ? 'z.string().datetime()'
@@ -102,10 +353,37 @@ function zodToString(schema: any, indentationLevel = 0): string {
   }
 }
 
+// Generate the string for all content type schemas and set tracker flags
+const contentTypeSchemasString = Object.entries(fieldSchemasMap)
+  .map(([key, fields]) => {
+    const fieldsSchemaString = zodToString(z.object(fields), 1);
+    return `
+// Schema for ${key}
+export const ${key}FieldsSchema = ${fieldsSchemaString};
+
+export const ${key}Schema = z.object({
+  metadata: metadataSchema,
+  sys: sysEntrySchema.extend({
+    contentType: z.object({
+      sys: z.object({
+        type: z.literal('Link'),
+        linkType: z.literal('ContentType'),
+        id: z.literal('${key}'),
+      }),
+    }),
+  }),
+  fields: ${key}FieldsSchema,
+});
+
+export type ${key} = z.infer<typeof ${key}Schema>;`;
+  })
+  .join('\n\n');
+
 // Convert schemas to TypeScript code
 const schemaCode = `import { z } from 'zod';
 import { BLOCKS } from '@contentful/rich-text-types';
 
+${usesFilterUnpublished ? filterUnpublishedHelper + '\n\n' : ''}${usesSingleReferenceTransform ? singleReferenceTransformHelper + '\n\n' : ''}
 // ========================================
 // Base Schemas
 // ========================================
@@ -199,33 +477,7 @@ export type RichTextNode = z.infer<typeof ${RICH_TEXT_SCHEMA_NAME}>;
 // Content Type Specific Schemas
 // ========================================
 
-${Object.entries(fieldSchemasMap)
-  .map(([key, fields]) => {
-    // Generate the string for the fields object schema using the simplified zodToString
-    const fieldsSchemaString = zodToString(z.object(fields), 1);
-
-    // Construct the full schema definition string for the content type
-    return `
-// Schema for ${key}
-export const ${key}FieldsSchema = ${fieldsSchemaString};
-
-export const ${key}Schema = z.object({
-  metadata: metadataSchema,
-  sys: sysEntrySchema.extend({
-    contentType: z.object({
-      sys: z.object({
-        type: z.literal('Link'),
-        linkType: z.literal('ContentType'),
-        id: z.literal('${key}'),
-      }),
-    }),
-  }),
-  fields: ${key}FieldsSchema,
-});
-
-export type ${key} = z.infer<typeof ${key}Schema>;`;
-  })
-  .join('\n\n')}
+${contentTypeSchemasString}
 
 // ========================================
 // Union Schema and Helper Object
